requirements.txt
python-dotenv
requests
psycopg2-binary
apache-airflow
dbt-core
dbt-postgres